feature: nacos 配置优化

+++ config/plugin/workbunny/webman-nacos/app.php
@@ -12,6 +12,7 @@ $groupName = get_env('NACOS_GROUP', 'DEFAULT_GROUP');
 $configListeners = collect([
     // 在此处增加配置中心的配置
     // $dateId => $filePath
+    // 'webman-env' => base_path() . '.env',
 ])->map(
     fn($filepath, $dataId) => [
         $dataId,
@@ -25,7 +26,7 @@ return [
     'enable' => true,
 
     'host' => get_env('NACOS_HOST', '127.0.0.1'),
-    'port' => get_env('NACOS_PORT', 8848),
+    'port' => (int)get_env('NACOS_PORT', 8848),
     /* username 和 password 同时设置为 null，不启用认证 */
     'username' => null,
     'password' => null,
@@ -58,7 +59,7 @@ return [
             get_env('NACOS_SERVICE_IP', Tools::getLocalIp()),
 
             /** port */
-            get_env('NACOS_SERVICE_PORT', Tools::getLocalServerPort()),
+            (int)get_env('NACOS_SERVICE_PORT', Tools::getLocalServerPort()),
 
             /** optional @see \Workbunny\WebmanNacos\Provider\InstanceProvider::registerAsync() */
             [
